Cart now stores and displays properly
Cart stores to localStorage now! The problem I had before was that it was resetting as the newest item and erasing the old entries.

Also, one of the issues I had with a console error was that it would complain whenever it couldn't find one of the spans I wanted to replace, so I changed it to also check the specific page the span is on (document.URL.includes("5cart.html").